Initialize tag colection in IdeaServiceImpl class, method findIdeasByTagId
