+++ src/main/java/com/epam/idea/core/service/impl/IdeaServiceImpl.java
@@ -82,7 +82,11 @@ public class IdeaServiceImpl implements IdeaService {
 
 	@Override
 	public List<Idea> findIdeasByTagId(final long tagId) {
-		return ideaRepository.findByTagId(tagId);
+		List<Idea> ideas = ideaRepository.findByTagId(tagId);
+		ideas.forEach(idea -> {
+			Hibernate.initialize(idea.getRelatedTags());
+		});
+		return ideas;
 	}
 
 	@Override
