+++ src/test/java/com/epam/idea/rest/controller/IdeaControllerTest.java
@@ -310,31 +310,4 @@ public class IdeaControllerTest {
 		assertThat(ideaArgument.getTitle()).isEqualTo(source.getTitle());
 		assertThat(ideaArgument.getDescription()).isEqualTo(source.getDescription());
 	}
-
-	@Test
-	public void shouldReturnAllFoundIdeasForGivenTag() throws Exception {
-		Tag tag = TestTagBuilder.aTag().build();
-		Idea idea = TestIdeaBuilder.anIdea().build();
-		tag.addIdea(idea);
-		idea.addTag(tag);
-
-		when(ideaServiceMock.findIdeasByTagId(tag.getId())).thenReturn(Lists.newArrayList(idea));
-		mockMvc.perform(get("/api/v1/tags/{tagId}/ideas", tag.getId())
-				.accept(APPLICATION_JSON_UTF8))
-				.andExpect(status().isOk())
-				.andExpect(content().contentType(APPLICATION_JSON_UTF8))
-				.andExpect(jsonPath("$", hasSize(1)))
-				.andExpect(jsonPath("$[0].title").value(is(idea.getTitle())))
-				.andExpect(jsonPath("$[0].description").value(is(idea.getDescription())))
-				.andExpect(jsonPath("$[0].rating").value(is(idea.getRating())))
-				.andExpect(jsonPath("$[0].tags", hasSize(1)))
-				.andExpect(jsonPath("$[0].tags[0]." + ID).value(is(((int) tag.getId()))))
-				.andExpect(jsonPath("$[0].tags[0].name").value(is(tag.getName())))
-				.andExpect(jsonPath("$[0].links", hasSize(1)))
-				.andExpect(jsonPath("$[0].links[0].rel").value(is(Link.REL_SELF)))
-				.andExpect(jsonPath("$[0].links[0].href").value(containsString("/api/v1/ideas/" + idea.getId())));
-
-		verify(ideaServiceMock, times(1)).findIdeasByTagId(tag.getId());
-		verifyNoMoreInteractions(ideaServiceMock);
-	}
 }
