+++ src/test/java/com/epam/idea/rest/controller/TagControllerTest.java
@@ -1,12 +1,16 @@
 package com.epam.idea.rest.controller;
 
+import com.epam.idea.core.model.Idea;
 import com.epam.idea.core.model.Tag;
+import com.epam.idea.core.model.builders.TestIdeaBuilder;
 import com.epam.idea.core.model.builders.TestTagBuilder;
+import com.epam.idea.core.service.IdeaService;
 import com.epam.idea.core.service.TagService;
 import com.epam.idea.core.service.exception.TagDoesNotExistException;
 import com.epam.idea.rest.config.TestConfig;
 import com.epam.idea.rest.config.WebAppConfig;
 import com.epam.idea.rest.resource.asm.TagResourceAsm;
+import com.google.common.collect.Lists;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -47,6 +51,9 @@ public class TagControllerTest {
 
 	@Autowired
 	private TagService tagServiceMock;
+
+	@Autowired
+	private IdeaService ideaServiceMock;
 
 	@Autowired
 	private WebApplicationContext webApplicationContext;
@@ -112,5 +119,37 @@ public class TagControllerTest {
 
 		verify(tagServiceMock, times(1)).findOne(TestTagBuilder.DEFAULT_ID);
 		verifyNoMoreInteractions(tagServiceMock);
+	}
+
+	@Test
+	public void shouldReturnAllFoundIdeasForGivenTag() throws Exception {
+		Tag tag = TestTagBuilder.aTag().build();
+		Idea idea = TestIdeaBuilder.anIdea().build();
+		tag.addIdea(idea);
+		idea.addTag(tag);
+
+		when(ideaServiceMock.findIdeasByTagId(tag.getId())).thenReturn(Lists.newArrayList(idea));
+		mockMvc.perform(get("/api/v1/tags/{tagId}/ideas", tag.getId())
+				.accept(APPLICATION_JSON_UTF8))
+				.andExpect(status().isOk())
+				.andExpect(content().contentType(APPLICATION_JSON_UTF8))
+				.andExpect(jsonPath("$", hasSize(1)))
+				.andExpect(jsonPath("$[0].title").value(is(idea.getTitle())))
+				.andExpect(jsonPath("$[0].description").value(is(idea.getDescription())))
+				.andExpect(jsonPath("$[0].rating").value(is(idea.getRating())))
+				.andExpect(jsonPath("$[0].tags", hasSize(1)))
+				.andExpect(jsonPath("$[0].tags[0]." + ID).value(is(((int) tag.getId()))))
+				.andExpect(jsonPath("$[0].tags[0].name").value(is(tag.getName())))
+				.andExpect(jsonPath("$[0].tags[0].links", hasSize(2)))
+				.andExpect(jsonPath("$[0].tags[0].links[0].rel").value(is(Link.REL_SELF)))
+				.andExpect(jsonPath("$[0].tags[0].links[0].href").value(containsString("/api/v1/tags/" + tag.getId())))
+				.andExpect(jsonPath("$[0].tags[0].links[1].rel").value(is(TagResourceAsm.IDEAS_REL)))
+				.andExpect(jsonPath("$[0].tags[0].links[1].href").value(containsString("/api/v1/tags/" + tag.getId() + "/ideas")))
+				.andExpect(jsonPath("$[0].links", hasSize(1)))
+				.andExpect(jsonPath("$[0].links[0].rel").value(is(Link.REL_SELF)))
+				.andExpect(jsonPath("$[0].links[0].href").value(containsString("/api/v1/ideas/" + idea.getId())));
+
+		verify(ideaServiceMock, times(1)).findIdeasByTagId(tag.getId());
+		verifyNoMoreInteractions(ideaServiceMock);
 	}
 }
